refactor: 상권 상세 정보 컴포넌트 렌더링 분리
