+++ apps/web/components/sidebar/DetailSideContents.tsx
@@ -1,60 +1,20 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { InfoBarData } from '../../types/map-types';
 import RevenueCard from './RevenueCard';
-import { MarketAnalysisData } from '@/types/market-types';
 import MarketBadge from './market/MarketBadge';
 import ReviewSummary from './market/ReviewSummary';
 import StoreList from './market/StoreList';
 import VitalityStats from './market/VitalityStats';
-import { createMarketAnalysisUrl } from '@/utils/map-utils';
 import DetailedStores from './market/DetailedStores';
+import { useMarketAnalysis } from '@/hooks/useMarketAnalysis';
 
 interface DetailContentsProps {
   data: InfoBarData;
 }
 
 export default function DetailContents({ data }: DetailContentsProps) {
-  const [analysisData, setAnalysisData] = useState<MarketAnalysisData | null>(null);
-  const [loading, setLoading] = useState(false);
+  const {analysisData, loading } = useMarketAnalysis(data);
   const [view, setView] = useState<'summary' | 'list'>('summary'); // 뷰 상태 관리
-
-  const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-
-  // ... (useEffect logic remains the same) ...
-  // But make sure to reset view when data changes
-  useEffect(() => {
-    setView('summary'); // 데이터 변경 시 뷰 초기화
-
-    const fetchMarketAnalysis = async () => {
-      setLoading(true);
-      setAnalysisData(null);
-
-      const finalUrl = createMarketAnalysisUrl(API_BASE_URL!, data.y, data.x, data.polygons);
-      console.log('최종 호출 URL:', finalUrl);
-
-      try {
-        console.log('Fetching analysis for:', data.y, data.x);
-        const res = await fetch(finalUrl);
-        
-        if (!res.ok) {
-           throw new Error('Network response was not ok');
-        }
-        const result: MarketAnalysisData = await res.json();
-        setAnalysisData(result);
-      } catch (error) {
-        console.error('분석 데이터를 가져오는데 실패함.', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-
-    if (data.x && data.y) {
-      fetchMarketAnalysis();
-    } else {
-      console.warn('좌표(x, y) 정보가 없어 상권 분석을 진행할 수 없습니다.', data);
-    }
-  }, [data, API_BASE_URL]);
 
   if (view === 'list' && analysisData) {
     return (
@@ -65,23 +25,16 @@ export default function DetailContents({ data }: DetailContentsProps) {
     );
   }
 
+  if(loading){
+    return <LoadingSpinner />;
+  }
+
+  if(!analysisData){
+    return <EmptyState />;
+  }
+
   return (
     <div className="flex-1 overflow-y-auto p-4 space-y-6">
-      {/* 1. 로딩 중 */}
-      {loading && (
-        <div className="flex flex-col items-center justify-center h-40 space-y-3">
-          <div className="w-8 h-8 border-4 border-blue-500 border-t-transparent rounded-full animate-spin"></div>
-          <p className="text-gray-500 text-sm">상권 분석 중...</p>
-        </div>
-      )}
-      {/* 2. 데이터 없음 (에러) */}
-      {!loading && !analysisData && (
-        <div className="text-center text-gray-500 py-10">
-          데이터를 불러올 수 없습니다.
-        </div>
-      )}
-      {/* 3. 데이터 있음 (렌더링) */}
-      {!loading && analysisData && (
         <>
           {/* 상권 뱃지 (Hot Place vs 일반) */}
           <MarketBadge
@@ -108,7 +61,23 @@ export default function DetailContents({ data }: DetailContentsProps) {
             closureRate={analysisData.closureRate}
           />
         </>
-      )}
+    </div>
+  );
+}
+
+function LoadingSpinner(){
+  return (
+    <div className="flex flex-col items-center justify-center h-40 space-y-3">
+      <div className="w-8 h-8 border-4 border-blue-500 border-t-transparent rounded-full animate-spin"></div>
+      <p className="text-gray-500 text-sm">상권 분석 중...</p>
+    </div>
+  );
+}
+
+function EmptyState(){
+  return (
+    <div className="text-center text-gray-500 py-10">
+      데이터를 불러올 수 없습니다.
     </div>
   );
 }
